Add "Open in GUI" option and -g/--gui
This is helpful if the computaion is more complicated than originally
anticipated, as it allows continuing it in a more advanced environment.
It also allows for better persistence, since menus usually cannot be
"unfocused", while the GUI program can.

The GUI program defaults to qalculate-gtk, or qalculate-qt, but can be
manually overriden with `--gui`. If none of these are availabe, the
option to open in GUI will be hidden from the user.